refactor: split terminal debug socket support
